requirements.txt
flask
spotipy
nltk
pandas
scikit-learn
joblib
python-dotenv